Actualizo atributos de entidades Viaje,Reserva,Usuario que estaban mal mapeadas y agrego la funcionalidad de actualizar la recaudación del conductor cuando finaliza un recorrido.

+++ src/main/java/com/tallerwebi/dominio/Conductor.java
@@ -24,7 +24,7 @@ public class Conductor {
     @Enumerated(EnumType.STRING)
     private TipoDeLicencia licencia;
     private float calificacion;
-    private float ganancia;
+    private Double ganancia;
     private boolean cuentaHabilitada;
     @Enumerated(EnumType.STRING)
     private EstadoConductor estadoConductor;
